+++ packages/neuron-wallet/src/block-sync-renderer/sync/indexer-connector.ts
@@ -115,18 +115,22 @@ export default class IndexerConnector {
     }
   }
 
+  private async initSync() {
+    await this.processNextBlockNumber()
+
+    while (this.pollingIndexer) {
+      const indexerTipBlock = await this.indexer.tip()
+      await this.synchronize(indexerTipBlock)
+      await CommonUtils.sleep(5000)
+    }
+  }
+
   public async connect() {
     try {
       this.indexer.startForever()
       this.pollingIndexer = true
 
-      await this.processNextBlockNumber()
-
-      while (this.pollingIndexer) {
-        const indexerTipBlock = await this.indexer.tip()
-        await this.synchronize(indexerTipBlock)
-        await CommonUtils.sleep(5000)
-      }
+      this.initSync()
     } catch (error) {
       logger.error(`Error connecting to Indexer: ${error.message}`)
       throw error