+++ packages/neuron-wallet/src/controllers/api.ts
@@ -23,7 +23,7 @@ import CustomizedAssetsController from './customized-assets'
 import SystemScriptInfo from 'models/system-script-info'
 import logger from 'utils/logger'
 import AssetAccountController from './asset-account'
-import { GenerateCreateAssetAccountTxParams, SendCreateAssetAccountTxParams, UpdateAssetAccountParams } from './asset-account'
+import { GenerateCreateAssetAccountTxParams, SendCreateAssetAccountTxParams, UpdateAssetAccountParams, MigrateACPParams } from './asset-account'
 import AnyoneCanPayController from './anyone-can-pay'
 import { GenerateAnyoneCanPayTxParams, GenerateAnyoneCanPayAllTxParams, SendAnyoneCanPayTxParams } from './anyone-can-pay'
 import { DeviceInfo, ExtendedPublicKey } from 'services/hardware/common'
@@ -63,6 +63,10 @@ export default class ApiController {
     if (command === 'delete-wallet' || command === 'backup-wallet') {
       // params: walletID
       this.walletsController.requestPassword(params, command)
+    }
+
+    if (command === 'migrate-acp') {
+      this.assetAccountController.showACPMigrationDialog(false)
     }
   }
 
@@ -418,6 +422,15 @@ export default class ApiController {
 
     handle("get-asset-account", async (_, params: { walletID: string, id: number }) => {
       return this.assetAccountController.getAccount(params)
+    })
+
+    handle('check-migrate-acp', async () => {
+      const allowMultipleOpen = true
+      return this.assetAccountController.showACPMigrationDialog(allowMultipleOpen)
+    })
+
+    handle('migrate-acp', async (_, params: MigrateACPParams) => {
+      return this.assetAccountController.migrateAcp(params)
     })
 
     handle('generate-send-to-anyone-can-pay-tx', async (_, params: GenerateAnyoneCanPayTxParams) => {