+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 0.33.2 (2020-11-26)
+
+[CKB v0.35.1](https://github.com/nervosnetwork/ckb/releases/tag/v0.35.1) was released on Sept. 14th, 2020. This version of CKB node is now bundled and preconfigured in Neuron. 
+
+### Hotfix
+
+* Upgrade Asset Account script config.
+
+### Bug fix
+
+* Prevent Neuron from quitting when the main window is closed on MacOS.
+
+
 # 0.34.0-rc1 (2020-11-21)
 
 This is a release candidate to preview the changes in the next official release and may not be stable. Welcome any questions or suggestions.